Extract oneOf and anyOf tests, make them unresolvable
