+++ test/specs/index.spec.js
@@ -491,472 +491,6 @@ describe('module', function () {
       });
     });
 
-    it('merges anyOf', function () {
-      const result = mergeAndTest({
-        allOf: [
-          {
-            anyOf: [
-              {
-                required: ['123']
-              }
-            ]
-          },
-          {
-            anyOf: [
-              {
-                required: ['123']
-              },
-              {
-                required: ['456']
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        anyOf: [
-          {
-            required: ['123']
-          },
-          {
-            required: ['123', '456']
-          }
-        ]
-      });
-    });
-
-    it('merges anyOf by finding valid combinations', function () {
-      const result = mergeAndTest({
-        allOf: [
-          {
-            anyOf: [
-              {
-                type: ['null', 'string', 'array']
-              },
-              {
-                type: ['null', 'string', 'object']
-              }
-            ]
-          },
-          {
-            anyOf: [
-              {
-                type: ['null', 'string']
-              },
-              {
-                type: ['integer', 'object', 'null']
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        anyOf: [
-          {
-            type: ['null', 'string']
-          },
-          {
-            type: 'null'
-          },
-          {
-            type: ['object', 'null']
-          }
-        ]
-      });
-    });
-
-    it.skip('extracts common logic', function () {
-      const result = mergeAndTest({
-        allOf: [
-          {
-            anyOf: [
-              {
-                type: ['null', 'string', 'array'],
-                minLength: 5
-              },
-              {
-                type: ['null', 'string', 'object'],
-                minLength: 5
-              }
-            ]
-          },
-          {
-            anyOf: [
-              {
-                type: ['null', 'string'],
-                minLength: 5
-              },
-              {
-                type: ['integer', 'object', 'null']
-              }
-            ]
-          }
-        ]
-      });
-
-      // TODO I think this is correct
-      // TODO implement functionality
-      expect(result).to.eql({
-        type: 'null',
-        minLength: 5,
-        anyOf: [
-          {
-            type: 'string'
-          }
-        ]
-      });
-    });
-
-    it.skip('merges anyOf into main schema if left with only one combination', function () {
-      const result = mergeAndTest({
-        required: ['abc'],
-        allOf: [
-          {
-            anyOf: [
-              {
-                required: ['123']
-              },
-              {
-                required: ['456']
-              }
-            ]
-          },
-          {
-            anyOf: [
-              {
-                required: ['123']
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        required: ['abc', '123']
-      });
-    });
-
-    it('merges nested allOf if inside singular anyOf', function () {
-      const result = mergeAndTest({
-        allOf: [
-          {
-            anyOf: [
-              {
-                required: ['123'],
-                allOf: [
-                  {
-                    required: ['768']
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            anyOf: [
-              {
-                required: ['123']
-              },
-              {
-                required: ['456']
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        anyOf: [
-          {
-            required: ['123', '768']
-          },
-          {
-            required: ['123', '456', '768']
-          }
-        ]
-      });
-    });
-
-    it('throws if no intersection at all', function () {
-      expect(function () {
-        mergeAndTest({
-          allOf: [
-            {
-              anyOf: [
-                {
-                  type: ['object', 'string', 'null']
-                }
-              ]
-            },
-            {
-              anyOf: [
-                {
-                  type: ['array', 'integer']
-                }
-              ]
-            }
-          ]
-        });
-      }).to.throw(/incompatible/);
-
-      expect(function () {
-        mergeAndTest({
-          allOf: [
-            {
-              anyOf: [
-                {
-                  type: ['object', 'string', 'null']
-                }
-              ]
-            },
-            {
-              anyOf: [
-                {
-                  type: ['array', 'integer']
-                }
-              ]
-            }
-          ]
-        });
-      }).to.throw(/incompatible/);
-    });
-
-    it('merges more complex oneOf', function () {
-      const result = merger({
-        allOf: [
-          {
-            oneOf: [
-              {
-                type: ['array', 'string', 'object'],
-                required: ['123']
-              },
-              {
-                required: ['abc']
-              }
-            ]
-          },
-          {
-            oneOf: [
-              {
-                type: ['string']
-              },
-              {
-                type: ['object', 'array'],
-                required: ['abc']
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        oneOf: [
-          {
-            type: 'string',
-            required: ['123']
-          },
-          {
-            type: ['object', 'array'],
-            required: ['123', 'abc']
-          },
-          {
-            type: ['string'],
-            required: ['abc']
-          },
-          {
-            type: ['object', 'array'],
-            required: ['abc']
-          }
-        ]
-      });
-    });
-
-    it('merges nested allOf if inside singular oneOf', function () {
-      const result = mergeAndTest({
-        allOf: [
-          {
-            type: ['array', 'string', 'number'],
-            oneOf: [
-              {
-                required: ['123'],
-                allOf: [
-                  {
-                    required: ['768']
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            type: ['array', 'string']
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        type: ['array', 'string'],
-        oneOf: [
-          {
-            required: ['123', '768']
-          }
-        ]
-      });
-    });
-
-    it('merges nested allOf if inside multiple oneOf', function () {
-      const result = merger({
-        allOf: [
-          {
-            type: ['array', 'string', 'number'],
-            oneOf: [
-              {
-                type: ['array', 'object'],
-                allOf: [
-                  {
-                    type: 'object'
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            type: ['array', 'string'],
-            oneOf: [
-              {
-                type: 'string'
-              },
-              {
-                type: 'object'
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        type: ['array', 'string'],
-        oneOf: [
-          {
-            type: 'object'
-          }
-        ]
-      });
-    });
-
-    it.skip('throws if no compatible when merging oneOf', function () {
-      expect(function () {
-        merger({
-          allOf: [
-            {},
-            {
-              oneOf: [
-                {
-                  required: ['123']
-                }
-              ]
-            },
-            {
-              oneOf: [
-                {
-                  required: ['fdasfd']
-                }
-              ]
-            }
-          ]
-        });
-      }).to.throw(/incompatible/);
-
-      expect(function () {
-        merger({
-          allOf: [
-            {},
-            {
-              oneOf: [
-                {
-                  required: ['123']
-                },
-                {
-                  properties: {
-                    name: {
-                      type: 'string'
-                    }
-                  }
-                }
-              ]
-            },
-            {
-              oneOf: [
-                {
-                  required: ['fdasfd']
-                }
-              ]
-            }
-          ]
-        });
-      }).to.throw(/incompatible/);
-    });
-
-    // not ready to implement this yet
-    it.skip('merges singular oneOf', function () {
-      const result = merger({
-        properties: {
-          name: {
-            type: 'string'
-          }
-        },
-        allOf: [
-          {
-            properties: {
-              name: {
-                type: 'string',
-                minLength: 10
-              }
-            }
-          },
-          {
-            oneOf: [
-              {
-                required: ['123']
-              },
-              {
-                properties: {
-                  name: {
-                    type: 'string',
-                    minLength: 15
-                  }
-                }
-              }
-            ]
-          },
-          {
-            oneOf: [
-              {
-                required: ['abc']
-              },
-              {
-                properties: {
-                  name: {
-                    type: 'string',
-                    minLength: 15
-                  }
-                }
-              }
-            ]
-          }
-        ]
-      });
-
-      expect(result).to.eql({
-        properties: {
-          name: {
-            type: 'string',
-            minLength: 15
-          }
-        }
-      });
-    });
-
     it('merges not using allOf', function () {
       const result = merger({
         allOf: [
@@ -1002,8 +536,6 @@ describe('module', function () {
         }
       });
     });
-
-    it('extracts pattern from anyOf and oneOf using | operator in regexp');
 
     it.skip('merges multipleOf using allOf or direct assignment', function () {
       const result = mergeAndTest({
